Create Encapsulation.Java
Encapsulation


You need a program to manage admissions for an art school. Pupils can be admitted to the school if they are over 6 years of age.
You're given a program which declares a Pupil class.

Task
Complete the setAge method of the Pupil class. If the value of parameter a is over 6, assign it to age attribute and output "Welcome".
Output "Sorry" otherwise.

Sample Input
7

Sample Output
Welcome
Use an if statement to check the mentioned condition.